backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.17.1
psycopg2-binary==2.9.12
python-dotenv==1.2.2
sqlparse==0.5.5
